files components: set up configurable file attributes
* refactor components
