+++ invenio_records_resources/services/records/components/__init__.py
@@ -1,0 +1,16 @@
+from .base import ServiceComponent, BaseRecordFilesComponent
+from .data import DataComponent
+from .metadata import MetadataComponent
+from .files import FilesOptionsComponent, AuxFilesOptionsComponent
+from .relations import RelationsComponent, ChangeNotificationsComponent
+
+__all__ = (
+    "ServiceComponent",
+    "DataComponent",
+    "MetadataComponent",
+    "FilesOptionsComponent",
+    "RelationsComponent",
+    "ChangeNotificationsComponent",
+    "BaseRecordFilesComponent",
+    "AuxFilesOptionsComponent",
+)
